move PendingBlocksAPI impl to pending_blocks.rs
