+++ crates/flashblocks-rpc/src/pending_blocks.rs
@@ -1,4 +1,6 @@
-use crate::subscription::Flashblock;
+use std::sync::Arc;
+
+use crate::{rpc::PendingBlocksAPI, subscription::Flashblock};
 use alloy_consensus::{Header, Sealed};
 use alloy_eips::BlockNumberOrTag;
 use alloy_primitives::{
@@ -8,11 +10,13 @@ use alloy_primitives::{
 use alloy_provider::network::TransactionResponse;
 use alloy_rpc_types::{state::StateOverride, BlockTransactions};
 use alloy_rpc_types_eth::{Filter, Header as RPCHeader, Log};
+use arc_swap::Guard;
 use eyre::eyre;
 use op_alloy_network::Optimism;
 use op_alloy_rpc_types::{OpTransactionReceipt, Transaction};
 use reth::revm::{db::Cache, state::EvmState};
-use reth_rpc_eth_api::RpcBlock;
+use reth_rpc_convert::RpcTransaction;
+use reth_rpc_eth_api::{RpcBlock, RpcReceipt};
 
 pub struct PendingBlocksBuilder {
     flashblocks: Vec<Flashblock>,
@@ -251,5 +255,54 @@ impl PendingBlocks {
         }
 
         logs
+    }
+}
+
+impl PendingBlocksAPI for Guard<Option<Arc<PendingBlocks>>> {
+    fn get_canonical_block_number(&self) -> BlockNumberOrTag {
+        self.as_ref()
+            .map(|pb| pb.canonical_block_number())
+            .unwrap_or(BlockNumberOrTag::Latest)
+    }
+
+    fn get_transaction_count(&self, address: Address) -> U256 {
+        self.as_ref()
+            .map(|pb| pb.get_transaction_count(address))
+            .unwrap_or_else(|| U256::from(0))
+    }
+
+    fn get_block(&self, full: bool) -> Option<RpcBlock<Optimism>> {
+        self.as_ref().map(|pb| pb.get_latest_block(full))
+    }
+
+    fn get_transaction_receipt(
+        &self,
+        tx_hash: alloy_primitives::TxHash,
+    ) -> Option<RpcReceipt<Optimism>> {
+        self.as_ref().and_then(|pb| pb.get_receipt(tx_hash))
+    }
+
+    fn get_transaction_by_hash(
+        &self,
+        tx_hash: alloy_primitives::TxHash,
+    ) -> Option<RpcTransaction<Optimism>> {
+        self.as_ref()
+            .and_then(|pb| pb.get_transaction_by_hash(tx_hash))
+    }
+
+    fn get_balance(&self, address: Address) -> Option<U256> {
+        self.as_ref().and_then(|pb| pb.get_balance(address))
+    }
+
+    fn get_state_overrides(&self) -> Option<StateOverride> {
+        self.as_ref()
+            .map(|pb| pb.get_state_overrides())
+            .unwrap_or_default()
+    }
+
+    fn get_pending_logs(&self, filter: &Filter) -> Vec<Log> {
+        self.as_ref()
+            .map(|pb| pb.get_pending_logs(filter))
+            .unwrap_or_default()
     }
 }
